refactor: escrow, ethereum endpoints.
