+++ src/app.module.ts
@@ -4,8 +4,8 @@ import { LabRating } from './endpoints/rating/models/rating.entity';
 import { LocationEntities } from './endpoints/location/models';
 import { LocationModule } from './endpoints/location/location.module';
 import { RatingModule } from './endpoints/rating/rating.module';
-import { EthereumModule } from './ethereum/ethereum.module';
-import { EscrowModule } from './escrow/escrow.module';
+import { EthereumModule } from './endpoints/ethereum/ethereum.module';
+import { EscrowModule } from './endpoints/escrow/escrow.module';
 import { SubstrateModule } from './substrate/substrate.module';
 import { SubstrateIndexedDataModule } from './substrate-indexed-data/substrate-indexed-data.module';
 import { EthereumIndexedDataModule } from './ethereum-indexed-data/ethereum-indexed-data.module';
