+++ src/endpoints/escrow/escrow.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { EscrowAccounts } from './models/deposit.entity';
 import { EscrowController } from './escrow.controller';
 import { EscrowService } from './escrow.service';
-import { SubstrateModule } from '../substrate/substrate.module';
+import { SubstrateModule } from '../../substrate/substrate.module';
 import { EthereumModule } from '../ethereum/ethereum.module';
 import { ElasticsearchModule } from '@nestjs/elasticsearch';
