+++ src/endpoints/escrow/escrow.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { formatEther, WalletSigner } from 'nestjs-ethers';
 import { EthereumService } from '../ethereum/ethereum.service';
-import { SubstrateService } from '../substrate/substrate.service';
+import { SubstrateService } from '../../substrate/substrate.service';
 import { Utils } from './utils/utils';
 import { ethers } from 'ethers';
 import ABI from '../ethereum/utils/ABI.json';
