+++ src/endpoints/ethereum/ethereum.service.spec.ts
