+++ src/ethereum-indexed-data/ethereum-indexed-data.module.ts
@@ -3,7 +3,7 @@ import { ElasticsearchModule } from '@nestjs/elasticsearch';
 import { ServiceRequestController } from './service-request/service-request.controller';
 import { ServiceRequestService } from './service-request/service-request.service';
 import { LocationModule } from '../endpoints/location/location.module';
-import { EthereumModule } from '../ethereum/ethereum.module';
+import { EthereumModule } from '../endpoints/ethereum/ethereum.module';
 import { DebioConversionModule } from 'src/debio-conversion/debio-conversion.module';
 
 @Module({
