+++ src/substrate/substrate.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { SubstrateService } from './substrate.service';
 import { SubstrateController } from './substrate.controller';
-import { EscrowModule } from '../escrow/escrow.module';
+import { EscrowModule } from '../endpoints/escrow/escrow.module';
 import { TransactionLoggingModule } from '../transaction-logging/transaction-logging.module';
 import { RewardModule } from 'src/reward/reward.module';
 import { MailModule } from 'src/common/mailer/mailer.module';
